Set up and initial code
I've done File Management setup, and made a basic code for a GameManager. Only thing needed is to figure out how to Drag and Drop, and trigger events on colliding objects, and things should be all groovy from there. Also, included support for ProBuilder, and 2D Plugins if we're switcing between a 2D and 3D Perspective in art style.